Migrate diktat smoke tests to SAVE-cli mechanism
### What's done:
* migrated diktat smoke tests to SAVE-cli mechanism
Closes #1383
